fix tests using new methods
