+++ test_vm/tests/verified_claim_test.rs
@@ -59,7 +59,7 @@ fn verified_claim_scenario() {
         seal_proof.registered_window_post_proof().unwrap(),
         &TokenAmount::from_whole(1_000),
     );
-    let mut v = v.with_epoch(200);
+    let v = v.with_epoch(200);
 
     // Register verifier and verified clients
     let datacap = StoragePower::from(32_u128 << 40);
@@ -104,7 +104,7 @@ fn verified_claim_scenario() {
     );
 
     // Advance time to max seal duration and prove the sector
-    v = advance_by_deadline_to_epoch(v, miner_id, deal_start).0;
+    advance_by_deadline_to_epoch(&v, &miner_id, deal_start);
     miner_prove_sector(&v, &worker, &miner_id, sector_number);
     // Trigger cron to validate the prove commit
     cron_tick(&v);
@@ -167,8 +167,8 @@ fn verified_claim_scenario() {
     );
 
     // Advance to proving period and submit post
-    let (deadline_info, partition_index, v) =
-        advance_to_proving_deadline(v, miner_id, sector_number);
+    let (deadline_info, partition_index) =
+        advance_to_proving_deadline(&v, &miner_id, sector_number);
 
     let expected_power =
         PowerPair { raw: StoragePower::from(deal_size), qa: StoragePower::from(10 * deal_size) };
@@ -188,17 +188,17 @@ fn verified_claim_scenario() {
     assert_eq!(power_claim.quality_adj_power, expected_power.qa);
 
     // move forward one deadline so advanceWhileProving doesn't fail double submitting posts.
-    let (mut v, _) = advance_by_deadline_to_index(
-        v,
-        miner_id,
+    advance_by_deadline_to_index(
+        &v,
+        &miner_id,
         deadline_info.index + 1 % policy.wpost_period_deadlines,
     );
 
     // Advance past the deal's minimum term (the claim remains valid).
-    v = advance_by_deadline_to_epoch_while_proving(
-        v,
-        miner_id,
-        worker,
+    advance_by_deadline_to_epoch_while_proving(
+        &v,
+        &miner_id,
+        &worker,
         sector_number,
         deal_start + deal_term_min + 10,
     );
@@ -222,10 +222,10 @@ fn verified_claim_scenario() {
     );
 
     // Advance toward the sector's expiration
-    v = advance_by_deadline_to_epoch_while_proving(
-        v,
-        miner_id,
-        worker,
+    advance_by_deadline_to_epoch_while_proving(
+        &v,
+        &miner_id,
+        &worker,
         sector_number,
         extended_expiration_1 - 100,
     );
@@ -253,10 +253,10 @@ fn verified_claim_scenario() {
     );
 
     // Advance toward the sector's new expiration
-    v = advance_by_deadline_to_epoch_while_proving(
-        v,
-        miner_id,
-        worker,
+    advance_by_deadline_to_epoch_while_proving(
+        &v,
+        &miner_id,
+        &worker,
         sector_number,
         extended_expiration_2 - 30 * EPOCHS_IN_DAY,
     );
@@ -306,10 +306,10 @@ fn verified_claim_scenario() {
     );
 
     // Advance sector to expiration
-    v = advance_by_deadline_to_epoch_while_proving(
-        v,
-        miner_id,
-        worker,
+    advance_by_deadline_to_epoch_while_proving(
+        &v,
+        &miner_id,
+        &worker,
         sector_number,
         extended_expiration_2,
     );
